git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@726 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/examples/event_filter/canvaspicker.cpp
@@ -214,7 +214,7 @@ void CanvasPicker::move(const QPoint &pos)
     QVector<double> xData(d_selectedCurve->dataSize());
     QVector<double> yData(d_selectedCurve->dataSize());
 
-    for ( int i = 0; i < d_selectedCurve->dataSize(); i++ )
+    for ( int i = 0; i < (int)d_selectedCurve->dataSize(); i++ )
     {
         if ( i == d_selectedPoint )
         {
